requirements.txt
rpyc
requests
beautifulsoup4
tldextract
sqlalchemy
cryptography
pymysql
gevent
grequests
dnspython